Create My Co-Parent App
It Takes A Village to Raise a Child:
Ever needed to quickly text someone like a co-parent when you're in a difficult parent moment? My Co-Parent is now here to provide quick advise on how to respond in a difficult parent moment with your child. 
Then give advise by submitting comments to the co-parent community and contribute to the village raising the child.